Rom: ReadByte function now takes uint instead of ushort
This allows accessing all data in larger roms. Ushort is too small for some potential rom sizes and requires a lot of added casting.

+++ Source/NEST/Classes/Rom.cs
@@ -124,7 +124,7 @@ namespace NEST.Classes
             }
         }
 
-        public byte readByte(ushort address)
+        public byte readByte(uint address)
         {
             return romData[address];
         }
